Use snaps for charm and charmcraft
Removes PyPI dependency for `charm` and `charmcraft` tools and instead
relies on them being installed at the system level, ideally as snaps.

In order to work with the confinement of the `charmcraft` snap, the temp
dirs have to be relocated to live under the Tox environment temp dir. It
also turns out that the `check_deps` fixture wasn't working as expected,
so it was fixed, as well as the need to call `charm build` without the
hyphen to support the snap.

As a bit of a drive-by, this also suppresses the DeprecationWarnings
generated by libjuju which are already being [tracked there][libjuju#461].

Fixes #13

[libjuju#461]: https://github.com/juju/python-libjuju/issues/461