Replace bespoke status with CLI call
The original intent was to avoid dependencies on the CLI, but due to
the complexity of the work done in the CLI vs the controller and the
gaps between that and libjuju, that ship has pretty much sailed. So,
instead of implementing a bespoke status output which differs from what
people expect, just use the CLI's existing status.