Add flag for destructive mode builds
Charmcraft builds charms in LXD containers by default, but there are
currently some bugs related to that and it's also less efficient if
you're already running in an ephemeral environment (e.g., a GH Workflow
runner). So this adds a flag to allow for skipping that by passing
through the `--destructive-mode` flag.