fix: clarify language in build_charm logging
This sets more explicit language for the log message emitted by `build_charm`.  Previously, if a charm build failed the user would see a message like:

    INFO     pytest_operator.plugin:plugin.py:591 Charm build for minio completed with return code 1 in 47.04s
    FAILED

This could be ambiguous as it implies the build succeeded in X seconds, when instead it finished (with error) in X seconds.  This commit clarifies that language