Add cos abstraction for testing grafana-agent-k8s
This is part of Kubeflow enhancement to integrate observability to
all charms (37). Since all tests will be the same in all charms, it
makes sense to put them on one place.
This help to test integration of any k8s charm and grafana-agent-k8s
charm, as it was discussed with observability team this is
recommended way to test such integration. Deploying whole cos should
not be necessary.

See the original implementation in Chisme [1].

---
[1]: https://github.com/canonical/charmed-kubeflow-chisme/pull/98